Typos (#150)
* Fixing the url for hello-world pipeline

Since the pipeline name is hello-world, the url should be http://127.0.0.1:8080/teams/main/pipelines/hello-world and not http://127.0.0.1:8080/teams/main/pipelines/helloworld

* fixing hello-world pipeline name and url

In the previous section, the pipeline name was  hello-world, and not  helloworld. To keep consistency, the name of the pipeline should be the same across different sections. 

I also fixed the hello-world url for the pipeline

*  fixing hello-world pipeline name

* fixing hello-world pipeline name

* fixing hello-world pipeline name

* fixing pipeline name