Remove 'mkdir' for dir already created by 'outputs'
In the 'basic/task-outputs-to-inputs' tutorial, the
`create_some_files.sh` attempts to first create the
output directory `some-files`.  However, since this
directory is also the name of one of the `outputs`,
it is already created implicitly.  This causes a
(harmless) error to appear in the task output.

Fix is to simply remove the `mkdir` command.