Added instructions to restart the Concourse server
Being new to both Concourse and Docker, it took me a bit to figure out I needed to restart Concourse in the local container after killing it at the end of the Introduction tutorial, and then a bit to figure out I had to do it from within the same directory.  Thought I'd try to help out other ignoramuses like myself.